River creation improved
No putting in a limit on how long a river can be, only need it to be accessible through grasslands. Takes a little while to load and hexmap class is quite long, may need refactoring and optimise (ugh)

Rivers are still not a hundred percent a straight line but that probably rivers aren't really in a straight line anyway.

Next is figuring out how to use sprites to create the look of the rivers